+++ test/test_api_v1_client_contact_company.rb
@@ -4,7 +4,7 @@
 require 'test_helper'
 
 class TestContactCompany < Minitest::Test
-  describe 'Contact Company Requests' do
+  describe 'Contact::Company Requests' do
     it 'get a contact company', :vcr do
       client = PapierkramApi::Client.new('simonneutert')
       response = client.contact_companies.by(id: 3)