+++ test/test_api_v1_client_tracker_tasks.rb
@@ -3,7 +3,7 @@
 require 'test_helper'
 
 class TestTrackerTasks < Minitest::Test
-  describe 'Tracker Tasks Requests' do
+  describe 'Tracker::Tasks Requests' do
     it 'gets a tracker tasks', :vcr do
       client = PapierkramApi::Client.new('simonneutert')
       response = client.tracker_tasks.by(id: 1)