+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [#22](https://github.com/simonneutert/papierkram_api_client/pull/22>) Factors out VcrSanitizer and adds some more Rubocop goodness. [@simonneutert](https://github.com/simonneutert)
 - [#21](https://github.com/simonneutert/papierkram_api_client/pull/21>) Adds more auto-sanitizing in VCR. [@simonneutert](https://github.com/simonneutert)
 
 - [#PR](https://github.com/simonneutert/papierkram_api_client/pull/<PRNUMBER>) Description of your PR contribution here. [@ghusername](https://github.com/<ghusername>)